requirements.txt
discord.py
sqlite3
Cybernator